grub.j2: fix grub-config-flags that use $GRUB_CMDLINE_LINUX_DEFAULT
The suggestion in README.md to 'add or keep kernel parameters you
had previously configured' does not work -- it's actually ignored.

This happens because grub-config-flags is evaluated in grub.j2
after we already set GRUB_CMDLINE_LINUX_DEFAULT, so it is gone.

Using it it grub-config-flags does not keep the previous value.

To fix that, move it to the first thing done in grub.j2, and
change our first assignment of GRUB_CMDLINE_LINUX_DEFAULT to
include it.

And now that our first assignment includes it, make sure it
is cleared up (ignored) if grub-config-flags is not set.

This is to keep the current behavior to ignore/override it.

Signed-off-by: Mauricio Faria de Oliveira <mfo@canonical.com>