functional/juju_tools: add method file_exists()
This could also be achieved by using the existing method file_stat()
using something like the following:

    import stat
    stat_result = await file_stat(...)
    assert stat.S_ISREG(stat_result)

The problem with this approach is that the `stat_result` object returned
by the remote python environment is being interpreted by the current
environment which might be prone to ABI and/or pickle version
incompatibilities.

Another problem with this approach is that in case of a
FileNotFoundError, the client side gets a very cryptic KeyError due to
"Stdout" key not existing in the result action.

The new file_exists() method calls os.path.isfile() on the remote
system and returns one of the python objects `True` or `False`, which
provides an arguably cleaner interface and is more expressive, i.e

    config_exists = await file_exists(...)
    assert config_exists

Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>