reactive/sysconfig.py: create sysctl config file upon installation
Currently, the sysctl config file in /etc/sysctl.d/ is only being
created during the "config-changed" event handler.

If sysctl config is given during `juju deploy`, i.e.

    $ juju deploy sysconfig --config sysctl="{ vm.swappiness:10 }"

`juju config` reports the value set by the operator but since the file
is not created and applied, the actual sysctl is not changed.

This patch fixes this behavior by creating the config file also during
the "install" handler, which will lead to the deploy-time settings
being applied.

Closes-Bug: #1915167
Signed-off-by: Mert Kırpıcı <mert.kirpici@canonical.com>